edit delete by quiz id, delete question in quiz

+++ src/services/quiz.service.js
@@ -19,7 +19,19 @@ const updateQuiz = async (id, data) => {
 }
 
 const deleteQuiz = async (id) => {
-    return await Quiz.findByIdAndDelete(id)
+    const quiz = await Quiz.findById(id)
+    if (!quiz) return null
+
+    const deleteQuestionsResult = await Question.deleteMany({
+        _id: { $in: quiz.questions }
+    })
+
+    await Quiz.findByIdAndDelete(id)
+
+    return {
+        quiz,
+        deletedQuestionsCount: deleteQuestionsResult.deletedCount
+    }
 }
 
 const getQuizWithQuestions = async(id) => {
